requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
pydantic[email]==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.20
starlette==0.45.3

# testing
pytest==8.3.4
httpx==0.28.1
pytest-mock==3.14.0

# lint
flake8==7.1.2
